fix: address cursor-plugin-cc UX gaps vs codex-plugin-cc
Tackles a batch of UX issues that put the plugin behind codex-plugin-cc:

- /cursor:result now defaults to the most recent terminal job for the
  workspace, matching codex's "chain the next thing" flow. The README
  quickstart no longer fails with UsageError.
- Status detail and result output now surface a copy-pasteable handoff
  block (`/cursor:resume`, `cursor-agent resume`, and the cursor.com agent
  URL) so users can move a thread into the standalone Cursor surface.
- disable-model-invocation: true is now set on every command (setup,
  task, review, adversarial-review). Stops Claude from re-invoking
  side-effecting commands through tool-use loops.
- renderJobTable, renderJobDetail, renderReviewResult, and renderReport
  emit Markdown so Claude Code's terminal renderer styles them.
  --json paths are unchanged.
- Job tables gain an Actions column with the suggested next slash
  command per row (/cursor:result <id> for terminal, /cursor:status +
  /cursor:cancel for active).
- Tables now distinguish elapsed (running, since startedAt) from
  duration (terminal, startedAt → finishedAt) and fall back to
  created-at age. JobIndexEntry carries startedAt/finishedAt to keep
  the table cheap.
- SessionEnd marks still-active jobs cancelled so closing Claude Code
  doesn't leave orphan agents burning Cursor usage.
  CURSOR_PLUGIN_KEEP_BACKGROUND_JOBS=1 opts out.
