+++ plugins/cursor/commands/result.md
@@ -1,5 +1,6 @@
 ---
-description: Print a completed job's result text.
+description: Print a completed job's result text. Defaults to the most recent terminal job in the current workspace.
+argument-hint: '[<job-id>] [--log] [--json]'
 allowed-tools: Bash(node:*)
 disable-model-invocation: true
 ---
@@ -10,10 +11,13 @@ disable-model-invocation: true
 node ${CLAUDE_PLUGIN_ROOT}/scripts/bundle/cursor-companion.mjs result $ARGUMENTS
 ```
 
-Required: a job id (from `/cursor:status`).
+Without an argument, prints the most recent terminal (`completed` / `failed` /
+`cancelled`) job for this workspace — the same shape codex-plugin-cc's
+`/codex:result` uses. Pass a job id (from `/cursor:status`) to target a
+specific job.
 
 Flags:
 - `--log` — print the streaming event log captured while the run was alive
 - `--json` — emit the full JobRecord
 
-If the job hasn't finished, exits 1. Surface the output verbatim.
+If the resolved job hasn't finished, exits 1. Surface the output verbatim.
