+++ tests/unit/lifecycle-hook.test.mts
@@ -60,4 +60,44 @@ describe("session-lifecycle-hook (vitest re-import)", () => {
     const sessionPath = path.join(stateDir, "session.json");
     expect(() => readFileSync(sessionPath)).toThrow();
   });
+
+  it("SessionEnd cancels still-active background jobs", async () => {
+    const { main } = await import("../../plugins/cursor/scripts/session-lifecycle-hook.mjs");
+    const { createJob, getJob, markRunning } = await import(
+      "../../plugins/cursor/scripts/lib/job-control.mjs"
+    );
+    const { ensureStateDir } = await import("../../plugins/cursor/scripts/lib/state.mjs");
+    const stateDir = ensureStateDir(resolveStateDir(resolveWorkspaceRoot(workDir)));
+
+    expect(main(["node", "hook", "SessionStart"])).toBe(0);
+    const running = createJob(stateDir, { type: "task", prompt: "still running" });
+    markRunning(stateDir, running.id, { agentId: "agent-r", runId: "run-r" });
+    const pending = createJob(stateDir, { type: "review", prompt: "still pending" });
+
+    expect(main(["node", "hook", "SessionEnd"])).toBe(0);
+
+    expect(getJob(stateDir, running.id)?.status).toBe("cancelled");
+    expect(getJob(stateDir, pending.id)?.status).toBe("cancelled");
+  });
+
+  it("SessionEnd respects CURSOR_PLUGIN_KEEP_BACKGROUND_JOBS=1", async () => {
+    const { main } = await import("../../plugins/cursor/scripts/session-lifecycle-hook.mjs");
+    const { createJob, getJob, markRunning } = await import(
+      "../../plugins/cursor/scripts/lib/job-control.mjs"
+    );
+    const { ensureStateDir } = await import("../../plugins/cursor/scripts/lib/state.mjs");
+    const stateDir = ensureStateDir(resolveStateDir(resolveWorkspaceRoot(workDir)));
+
+    expect(main(["node", "hook", "SessionStart"])).toBe(0);
+    const running = createJob(stateDir, { type: "task", prompt: "still running" });
+    markRunning(stateDir, running.id, { agentId: "agent-k", runId: "run-k" });
+
+    process.env.CURSOR_PLUGIN_KEEP_BACKGROUND_JOBS = "1";
+    try {
+      expect(main(["node", "hook", "SessionEnd"])).toBe(0);
+    } finally {
+      delete process.env.CURSOR_PLUGIN_KEEP_BACKGROUND_JOBS;
+    }
+    expect(getJob(stateDir, running.id)?.status).toBe("running");
+  });
 });
